requirements.txt
# Hampi Monument Identifier — T12.5
# Python 3.10+

# ── Core ML ──────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0          # HuggingFace CLIP
Pillow>=10.0.0
numpy>=1.24.0

# ── Streamlit frontend ───────────────────────
streamlit>=1.30.0

# ── Optional: progress & utilities ──────────
tqdm>=4.66.0
requests>=2.31.0              # for future Wikimedia API calls

# ── Notebook (optional) ──────────────────────
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
